--- src/app/(app)/api/hello/route.ts
@@ -1,5 +1,0 @@
-export const dynamic = "force-static";
-
-export async function GET() {
-  return Response.json({ data: { hi: "there" } });
-}
